restructures resources model file
