+++ resources/models/__init__.py
@@ -1,0 +1,13 @@
+from resources.models.home import Home, FormField
+from resources.models.resources import ResourceIndexPage, ResourcePage
+from resources.models.tags import IssueTag, TopicTag, ContentTag
+
+__all__ = [
+    'Home',
+    'FormField',
+    'ResourceIndexPage',
+    'ResourcePage',
+    'IssueTag',
+    'TopicTag',
+    'ContentTag'
+]
