+++ resources/models/home.py
@@ -1,37 +1,23 @@
-from __future__ import unicode_literals
-
-from urllib.parse import parse_qs
-
-from django.db import models
-from django.db.models import Q, Sum, Case, When
-from django.db.models.fields import TextField, URLField, IntegerField, CharField
-from django.shortcuts import render
-
-from django.contrib import messages
-from django.contrib.messages import get_messages
-from django.http import HttpResponseRedirect
-
-from wagtail.wagtailcore.models import Page
+from wagtail.wagtailforms.models import AbstractForm, AbstractFormField
 from wagtail.wagtailcore.fields import RichTextField
 from wagtail.wagtailadmin.edit_handlers import FieldPanel, InlinePanel
-from modelcluster.fields import ParentalKey
-from modelcluster.contrib.taggit import ClusterTaggableManager
-from taggit.models import TaggedItemBase, Tag
-from wagtail.wagtailsearch import index
-
-from wagtail.wagtailimages.models import Image
 from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 
-from wagtail.wagtailforms.models import AbstractForm, AbstractFormField
+from django.db.models.fields import TextField, URLField
+from django.db import models
+from django.db.models import Q
 
-from django.template.loader import get_template
+from django.contrib.messages import get_messages
+from django.shortcuts import render
 
+from modelcluster.fields import ParentalKey
+from urllib.parse import parse_qs
 from itertools import chain
 
-from likes.models import Likes
+from resources.models.tags import TopicTag, IssueTag, ReasonTag, ContentTag
+from resources.models.resources import ResourcePage
+from resources.models.helpers import combine_tags, count_likes, get_liked_value, filter_tags, get_tags
 
-import queue
-import threading, time, random
 
 class FormField(AbstractFormField):
     page = ParentalKey('Home', related_name='form_fields')
@@ -246,181 +232,3 @@ class Home(AbstractForm):
             self.get_template(request),
             context
         )
-
-class TopicTag(TaggedItemBase):
-    content_object = ParentalKey('resources.ResourcePage', related_name='tagged_topic_items')
-
-class IssueTag(TaggedItemBase):
-    content_object = ParentalKey('resources.ResourcePage', related_name='tagged_issue_items')
-
-class ReasonTag(TaggedItemBase):
-    content_object = ParentalKey('resources.ResourcePage', related_name='tagged_reason_items')
-
-class ContentTag(TaggedItemBase):
-    content_object = ParentalKey('resources.ResourcePage', related_name='tagged_content_items')
-
-class HiddenTag(TaggedItemBase):
-    content_object = ParentalKey('resources.ResourcePage', related_name='tagged_hidden_items')
-
-class ResourceIndexPage(Page):
-    intro = RichTextField(blank=True)
-
-    def get_context(self, request):
-        context = super(ResourceIndexPage, self).get_context(request)
-        resources = self.get_children().live().order_by('-first_published_at')
-        context['resources'] = resources
-        return context
-
-    content_panels = Page.content_panels + [
-        FieldPanel('intro', classname="full")
-    ]
-
-class ResourcePage(Page):
-    heading = TextField(blank=True, help_text="The title of the resource being linked to")
-    resource_url = URLField(blank=True, help_text="The url of the resource to link to")
-    body = RichTextField(blank=True, help_text="A description of the resource")
-    pros = RichTextField(blank=True, help_text="A list of pros for the resource")
-    cons = RichTextField(blank=True, help_text="A list of cons for the resource")
-    video_url = URLField(blank=True, help_text="URL of a youtube video for the resource")
-
-    topic_tags = ClusterTaggableManager(
-        through=TopicTag, blank=True,
-        verbose_name='Topic Tags', related_name='resource_topic_tags',
-        help_text='Topic tags, eg: "sleep", "depression", "stress"'
-    )
-    issue_tags = ClusterTaggableManager(
-        through=IssueTag, blank=True,
-        verbose_name='Issue Tags', related_name='resource_issue_tags',
-        help_text='Issue tags, eg: "insomnia", "fatigue", "snoring"'
-    )
-    reason_tags = ClusterTaggableManager(
-        through=ReasonTag, blank=True,
-        verbose_name='Reason Tags', related_name='resource_reason_tags',
-        help_text='Reason tags, eg: "loneliness", "relationships"'
-    )
-    content_tags = ClusterTaggableManager(
-        through=ContentTag, blank=True,
-        verbose_name='Content Tags', related_name='resource_content_tags',
-        help_text='Content Type tags, eg: "videos", "blogs", "free", "subscription"'
-    )
-    hidden_tags = ClusterTaggableManager(
-        through=HiddenTag, blank=True,
-        verbose_name='Hidden Tags', related_name='resource_hidden_tags',
-        help_text='Hidden tags for admin use'
-    )
-    PRIORITY_CHOICES = (
-      (1, '1'),
-      (2, '2'),
-      (3, '3'),
-      (4, '4'),
-      (5, '5'),
-    )
-    priority = IntegerField(
-      choices=PRIORITY_CHOICES,
-      default='5',
-      help_text='Highest priority 1, lowest priority 5'
-    )
-
-    search_fields = Page.search_fields + [
-        index.SearchField('body'),
-        index.SearchField('pros'),
-        index.SearchField('cons'),
-        index.SearchField('heading'),
-        index.SearchField('resource_url'),
-        index.RelatedFields('issue_tags', [
-            index.SearchField('name'),
-        ]),
-        index.RelatedFields('content_tags', [
-            index.SearchField('name'),
-        ]),
-        index.RelatedFields('reason_tags', [
-            index.SearchField('name'),
-        ]),
-    ]
-
-    content_panels = Page.content_panels + [
-        FieldPanel('heading', classname="full"),
-        FieldPanel('resource_url', classname="full"),
-        FieldPanel('body', classname="full"),
-        FieldPanel('pros', classname="full"),
-        FieldPanel('cons', classname="full"),
-        FieldPanel('video_url', classname="full"),
-    ]
-
-    promote_panels = Page.promote_panels + [
-        FieldPanel('topic_tags'),
-        FieldPanel('issue_tags'),
-        FieldPanel('reason_tags'),
-        FieldPanel('content_tags'),
-        FieldPanel('hidden_tags'),
-        FieldPanel('priority'),
-    ]
-
-    class Meta:
-        verbose_name = "Resource"
-
-def get_tags(tag_type, **kwargs):
-    filtered_tags = kwargs.get('filtered_tags', tag_type.objects.all())
-    tag_ids = [tag.tag_id for tag in filtered_tags]
-    tags = Tag.objects.in_bulk(tag_ids)
-
-    return tags
-
-def combine_tags(element):
-    element.specific.tags = list(chain(
-        element.specific.content_tags.all(),
-        element.specific.issue_tags.all(),
-        element.specific.reason_tags.all(),
-    ))
-    return element
-
-def get_resource(id, user_hash):
-    return combine_tags(
-        ResourcePage.objects
-        .annotate(number_of_likes=count_likes(1))
-        .annotate(number_of_dislikes=count_likes(-1))
-        .annotate(liked_value=get_liked_value(user_hash))
-        .get(id=id)
-    )
-
-def count_likes(like_or_dislike):
-    return Sum(
-        Case(
-            When(likes__like_value=like_or_dislike, then=1),
-            default=0,
-            output_field=models.IntegerField()
-        )
-    )
-
-def get_liked_value(user_hash):
-    return Case(
-        When(likes__user_hash=user_hash, then='likes__like_value'),
-        default=0,
-        output_field=models.IntegerField()
-    )
-
-def filter_tags(resources, topic):
-    contains_tag_filter = makefilter(topic)
-
-    fil = list(map(lambda r: r.id, filter(contains_tag_filter, resources)))
-    q = queue.Queue()
-
-    threads = [ threading.Thread(target=filter_resource_by_topic, name=t, args=(fil, t, q)) for t in [ContentTag, IssueTag, ReasonTag] ]
-    # Using threads to make sure all db calls are completed before function returns
-
-    for th in threads:
-        th.start()
-
-    content_tags = q.get()
-    issue_tags = q.get()
-    reason_tags = q.get()
-
-    return issue_tags, reason_tags, content_tags
-
-def makefilter(t):
-    def contains_tag(r):
-        return any(filter(lambda tag: tag.name == t, r.specific.topic_tags.all()))
-    return contains_tag
-
-def filter_resource_by_topic(resource_ids, tag_type, result_queue):
-    result_queue.put(tag_type.objects.filter(content_object__in=resource_ids))
